Delete unnecessary calculation
When changing the default Min and/or Max values (Min=0, Max=100) for a dimmer block inside Loxone Config, and you start moving the slider within OpenHAB or Loxone app, strange values appear. I pinpointed the issue to be these two additional calculations. Therefore I would propose to delete the calculations and let the Min/Max value to be freely chosen from Loxone Config itself. This also has an additional advantage that someone is able to chose a range from 0-500 for example is he/she likes that.